+++ CHANGELOG.md
@@ -14,6 +14,8 @@ also specify classes on headers now
 - `zola serve/build` can now run from anywhere in a zola directory
 - Add XML support to `load_data`
 - `skip_prefixes` is now checked before parsing external link URLs
+- Add `render` attribute to taxonomies configuration in `config.toml`, for when you don't want to render
+any pages related to that taxonomy
 
 ## 0.15.3 (2022-01-23)